Make jail -k try harder to cleanup any messes left behind.
This removes the requirement that the mounts actually have directories.
It is possible that there are mounts still around but they are
removed somehow.  This can happen if the jail root is a ZFS
dataset but has nullfs mounts inside of it.  ZFS seems to allow
destroying the dataset even if it has mounts inside of it, thus
poudriere will never unmount them.  This was seen in Issue #296
with an MDMFS bug.

+++ src/share/poudriere/include/fs.sh
@@ -159,8 +159,7 @@ umountfs() {
 
 	[ -n "${childonly}" ] && pattern="/"
 
-	[ -d "${mnt}" ] || return 0
-	mnt=$(realpath ${mnt})
+	mnt=$(realpath "${mnt}" 2>/dev/null || echo "${mnt}")
 	xargsmax=
 	if [ ${UMOUNT_BATCHING} -eq 0 ]; then
 		xargsmax="-n 2"
@@ -177,7 +176,7 @@ _zfs_getfs() {
 	[ $# -ne 1 ] && eargs _zfs_getfs mnt
 	local mnt="${1}"
 
-	mntres=$(realpath "${mnt}")
+	mntres=$(realpath "${mnt}" 2>/dev/null || echo "${mnt}")
 	zfs list -rt filesystem -H -o name,mountpoint ${ZPOOL}${ZROOTFS} | \
 	    awk -vmnt="${mntres}" '$2 == mnt {print $1}'
 }
@@ -278,7 +277,7 @@ destroyfs() {
 	local mnt fs type
 	mnt=$1
 	type=$2
-	[ -d ${mnt} ] || return 0
+
 	umountfs ${mnt} 1
 	if [ ${TMPFS_ALL} -eq 1 ]; then
 		if [ -d "${mnt}" ]; then
@@ -290,10 +289,11 @@ destroyfs() {
 		[ "${fs}" != "none" ] && fs=$(zfs_getfs ${mnt})
 		if [ -n "${fs}" -a "${fs}" != "none" ]; then
 			zfs destroy -rf ${fs}
-			rmdir ${mnt}
+			rmdir ${mnt} || :
 			# Must invalidate the zfs_getfs cache.
 			cache_invalidate _zfs_getfs "${mnt}"
 		else
+			[ -d ${mnt} ] || return 0
 			rm -rfx ${mnt} 2>/dev/null || :
 			if [ -d "${mnt}" ]; then
 				chflags -R 0 ${mnt}
